refactor(runners): consolidate provider registry
