--- lambdas/functions/control-plane/src/pool/pool-provider-registry.ts
@@ -1,13 +1,0 @@
-import type { RunnerProviderType } from '../runner-provider';
-import { createEc2PoolProvider } from './ec2-pool';
-import type { PoolRunnerProvider } from './pool-provider';
-
-type PoolRunnerProviderFactory = () => Omit<PoolRunnerProvider, 'type'>;
-
-const poolRunnerProviderFactories: Record<RunnerProviderType, PoolRunnerProviderFactory> = {
-  ec2: createEc2PoolProvider,
-};
-
-export function createPoolRunnerProvider(type: RunnerProviderType): PoolRunnerProvider {
-  return { ...poolRunnerProviderFactories[type](), type };
-}
