+++ lambdas/functions/control-plane/src/pool/pool.ts
@@ -3,9 +3,9 @@ import { createChildLogger } from '@aws-github-runner/aws-powertools-util';
 import yn from 'yn';
 
 import { createGithubAppAuth, createGithubInstallationAuth, createOctokitClient } from '../github/auth';
+import { createPoolRunnerProvider } from '../runner-provider-registry';
 import { resolveRunnerProviderType } from '../runner-provider';
 import { getGitHubEnterpriseApiUrl, validateSsmParameterStoreTags } from '../scale-runners/github-runner';
-import { createPoolRunnerProvider } from './pool-provider-registry';
 import type { RunnerStatus } from './pool-provider';
 
 const logger = createChildLogger('pool');
