--- lambdas/functions/control-plane/src/scale-runners/scale-up-provider-registry.ts
@@ -1,13 +1,0 @@
-import type { RunnerProviderType } from '../runner-provider';
-import { createEc2ScaleUpProvider } from './ec2-scale-up';
-import type { ScaleUpRunnerProvider } from './scale-up-provider';
-
-type ScaleUpRunnerProviderFactory = () => Omit<ScaleUpRunnerProvider, 'type'>;
-
-const scaleUpRunnerProviderFactories: Record<RunnerProviderType, ScaleUpRunnerProviderFactory> = {
-  ec2: createEc2ScaleUpProvider,
-};
-
-export function createScaleUpRunnerProvider(type: RunnerProviderType): ScaleUpRunnerProvider {
-  return { ...scaleUpRunnerProviderFactories[type](), type };
-}
